pyproject.toml
[project]
name = "youtube-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper>=20250625",
    "pytube>=15.0.0",
    "torch==2.0.1+cu118",
    "torchaudio==2.0.2+cu118",
    "torchvision==0.15.2+cu118",
    "yt-dlp>=2025.9.5",
]
